Move serializeProcessing to utils
