+++ packages/edit-navigation/src/store/actions.js
@@ -14,15 +14,12 @@ import { serialize } from '@wordpress/blocks';
  * Internal dependencies
  */
 import {
-	getNavigationPostForMenu,
-	getPendingActions,
-	isProcessingPost,
 	getMenuItemToClientIdMapping,
 	resolveMenuItems,
 	dispatch,
 	apiFetch,
 } from './controls';
-import { menuItemsQuery } from './utils';
+import { menuItemsQuery, serializeProcessing } from './utils';
 
 /**
  * Creates a menu item for every block that doesn't have an associated menuItem.
@@ -228,62 +225,4 @@ function computeCustomizedAttribute( blocks, menuId, menuItemsByClientId ) {
 	function getMenuItemForBlock( block ) {
 		return omit( menuItemsByClientId[ block.clientId ] || {}, '_links' );
 	}
-}
-
-/**
- * This wrapper guarantees serial execution of data processing actions.
- *
- * Examples:
- * * saveNavigationPost() needs to wait for all the missing items to be created.
- * * Concurrent createMissingMenuItems() could result in sending more requests than required.
- *
- * @param {Function} callback An action creator to wrap
- * @return {Function} Original callback wrapped in a serial execution context
- */
-function serializeProcessing( callback ) {
-	return function* ( post ) {
-		const postId = post.id;
-		const isProcessing = yield isProcessingPost( postId );
-
-		if ( isProcessing ) {
-			yield {
-				type: 'ENQUEUE_AFTER_PROCESSING',
-				postId,
-				action: callback,
-			};
-			return { status: 'pending' };
-		}
-		yield {
-			type: 'POP_PENDING_ACTION',
-			postId,
-			action: callback,
-		};
-
-		yield {
-			type: 'START_PROCESSING_POST',
-			postId,
-		};
-
-		try {
-			yield* callback(
-				// re-select the post as it could be outdated by now
-				yield getNavigationPostForMenu( post.meta.menuId )
-			);
-		} finally {
-			yield {
-				type: 'FINISH_PROCESSING_POST',
-				postId,
-				action: callback,
-			};
-
-			const pendingActions = yield getPendingActions( postId );
-			if ( pendingActions.length ) {
-				const serializedCallback = serializeProcessing(
-					pendingActions[ 0 ]
-				);
-
-				yield* serializedCallback( post );
-			}
-		}
-	};
 }
